AkaServer/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.1
mssql-django==1.5
django-cors-headers==4.3.1
django-environ==0.11.2
wheel==0.45.0
environ==1.0
